Support assimp models (#1)
* Assimp usage
* added model outline using stencil test
* typo
* minor code refactors in other subdirectories

+++ core/camera.cpp
@@ -47,8 +47,8 @@ void Camera::UpdateYawPitchCallback(float xpos, float ypos) {
     firstTime = false;
   }
 
-  float xoffset = xpos - lastX;
-  float yoffset = lastY - ypos;
+  float xoffset = static_cast<float>(xpos - lastX);
+  float yoffset = static_cast<float>(lastY - ypos);
   lastX = xpos;
   lastY = ypos;
